添加一个bean和dao类，添加了每月还款并存储信息，保证数据库的及时更新，包括等额本金的提前还款

+++ demo/src/test/java/com/example/demo/Dao/auditMapper.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Mapper
 public interface auditMapper {
     @Select("select * from audit where checkNum = #{checkNum}")
-    public Audit selectAudioByCheckNum(int checkNum);
+    public Audit selectAudioByCheckNum(long checkNum);
 
     @Select("select * from audit")
     public List<Audit> select();
@@ -31,5 +31,5 @@ public interface auditMapper {
     public void updateAudit(Audit audits);
 
     @Delete("delete from audit where checkNum=#{checkNum}")
-    public void deleteAudit(int checkNum);
+    public void deleteAudit(long checkNum);
 }
